requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
Pillow
python-docx
googletrans==4.0.0-rc1
jinja2
python-multipart
